requirements.txt
absl-py==2.2.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chumpy @ git+https://github.com/mattloper/chumpy.git@51d5afd92a8ded3637553be8cef41f328a1c863a
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
Cython==3.1.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
diff_gaussian_rasterization @ file:///local/home/idemir/Desktop/3dv_dynamichumansplatting/InstantSplat/submodules/diff-gaussian-rasterization
diff_gaussian_rasterization_depth @ file:///local/home/idemir/Desktop/3dv_dynamichumansplatting/instantsplat/diff-gaussian-rasterization-depth
einops==0.8.1
evo==1.31.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.0
fused_ssim @ file:///local/home/idemir/Desktop/3dv_dynamichumansplatting/InstantSplat/submodules/fused-ssim
fvcore==0.1.5.post20221221
gdown==5.2.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gradio==5.22.0
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
icecream==2.1.4
idna @ file:///croot/idna_1714398848350/work
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iopath==0.1.10
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
kiwisolver==1.4.8
lazy_loader==0.4
libigl==2.5.0
loguru==0.7.2
lpips==0.1.4
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
multidict==6.4.3
narwhals==1.28.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
orjson==3.10.15
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plyfile==1.0.3
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
pydub==0.25.1
pyglet==1.5.31
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
-e git+ssh://git@github.com/iremddemir/3dv_dynamichumansplatting.git@01062c12b635f48d7343523c3b026111d4bea014#egg=pytorch3d&subdirectory=pytorch3d
pytz==2025.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
referencing==0.36.2
requests @ file:///croot/requests_1730999120400/work
rerun-sdk==0.22.1
retrying==1.3.4
rich==13.9.4
roma==1.5.2.1
rosbags==0.10.9
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
simple_knn @ file:///local/home/idemir/Desktop/3dv_dynamichumansplatting/InstantSplat/submodules/simple-knn
six==1.17.0
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.3.30
tomlkit==0.13.2
torch==2.3.1
torch-geometric==2.6.1
torchaudio==2.1.2
torchvision==0.18.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.1.7
triton==2.3.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
